requirements.txt
pycryptodome
sympy
cryptography